fix: tighten display semantic ownership
