+++ src/config/sessions/session-transcript-display.ts
@@ -1,3 +1,4 @@
+export * from "./session-transcript-display-reducer-contract.js";
 export * from "./session-transcript-display-semantics.js";
 export * from "./session-transcript-display-store.js";
 export * from "./session-transcript-display-rebuild-store.js";
